Convert device-view.js to typescript
