+++ cypress/integration/functionality.js
@@ -1,6 +1,5 @@
 import {runFrontendTest} from '../support/utils';
 
-
 describe('Motd testing', () => {
     before(() => {
         cy.visit('/');
